Broke out configuration validation from the mapper configuration
